Main game Tetris

+++ app/src/main/java/com/fiz/tetriswithlife/Cell.kt
@@ -1,5 +1,5 @@
 package com.fiz.tetriswithlife
 
-class Cell(x:Double,y:Double,val view:Int):Point(x,y) {
+class Cell(x:Float,y:Float,val view:Int):Point(x,y) {
 }
